# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateCompareRequest": {
        "additionalProperties": false,
        "properties": {
          "candidateSectionIds": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Candidatesectionids",
            "type": "array"
          },
          "currentSectionIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Currentsectionids",
            "type": "array"
          }
        },
        "required": [
          "candidateSectionIds"
        ],
        "title": "CandidateCompareRequest",
        "type": "object"
      },
      "CandidateCompareResponse": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateComparisonRead"
            },
            "title": "Candidates",
            "type": "array"
          }
        },
        "required": [
          "candidates"
        ],
        "title": "CandidateCompareResponse",
        "type": "object"
      },
      "CandidateComparisonRead": {
        "additionalProperties": false,
        "properties": {
          "added": {
            "items": {
              "type": "string"
            },
            "title": "Added",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "metrics": {
            "$ref": "#/components/schemas/CandidateMetrics"
          },
          "rank": {
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "removed": {
            "items": {
              "type": "string"
            },
            "title": "Removed",
            "type": "array"
          },
          "sectionIds": {
            "items": {
              "type": "string"
            },
            "title": "Sectionids",
            "type": "array"
          },
          "swapped": {
            "items": {
              "$ref": "#/components/schemas/CandidateSwapRead"
            },
            "title": "Swapped",
            "type": "array"
          }
        },
        "required": [
          "rank",
          "sectionIds",
          "metrics",
          "added",
          "removed",
          "swapped",
          "conflicts"
        ],
        "title": "CandidateComparisonRead",
        "type": "object"
      },
      "CandidateMetrics": {
        "additionalProperties": false,
        "properties": {
          "campusDays": {
            "title": "Campusdays",
            "type": "integer"
          },
          "firstClassMinute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstclassminute"
          },
          "gapMinutes": {
            "title": "Gapminutes",
            "type": "integer"
          },
          "lastClassMinute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastclassminute"
          },
          "targetCreditDeviation": {
            "default": 0,
            "title": "Targetcreditdeviation",
            "type": "number"
          },
          "totalCredits": {
            "title": "Totalcredits",
            "type": "number"
          },
          "unknownTimeSections": {
            "default": 0,
            "title": "Unknowntimesections",
            "type": "integer"
          }
        },
        "required": [
          "totalCredits",
          "campusDays",
          "gapMinutes",
          "firstClassMinute",
          "lastClassMinute"
        ],
        "title": "CandidateMetrics",
        "type": "object"
      },
      "CandidateSwapRead": {
        "additionalProperties": false,
        "properties": {
          "fromSectionId": {
            "title": "Fromsectionid",
            "type": "string"
          },
          "toSectionId": {
            "title": "Tosectionid",
            "type": "string"
          }
        },
        "required": [
          "fromSectionId",
          "toSectionId"
        ],
        "title": "CandidateSwapRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptimizationCandidate": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Changes",
            "type": "array"
          },
          "explanation": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Explanation",
            "type": "array"
          },
          "metrics": {
            "$ref": "#/components/schemas/CandidateMetrics"
          },
          "rank": {
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "scoreComponents": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Scorecomponents",
            "type": "object"
          },
          "sectionIds": {
            "items": {
              "type": "string"
            },
            "title": "Sectionids",
            "type": "array"
          },
          "unmetPreferences": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Unmetpreferences",
            "type": "array"
          }
        },
        "required": [
          "rank",
          "sectionIds",
          "metrics",
          "scoreComponents"
        ],
        "title": "OptimizationCandidate",
        "type": "object"
      },
      "OptimizationCreate": {
        "additionalProperties": false,
        "properties": {
          "candidateCount": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Candidatecount",
            "type": "integer"
          },
          "candidateCourseCodes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Candidatecoursecodes",
            "type": "array"
          },
          "datasetVersion": {
            "title": "Datasetversion",
            "type": "string"
          },
          "excludedCourseCodes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Excludedcoursecodes",
            "type": "array"
          },
          "lockedSectionIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Lockedsectionids",
            "type": "array"
          },
          "maxCredits": {
            "default": 18,
            "maximum": 30.0,
            "minimum": 0.0,
            "title": "Maxcredits",
            "type": "integer"
          },
          "minCredits": {
            "default": 12,
            "maximum": 30.0,
            "minimum": 0.0,
            "title": "Mincredits",
            "type": "integer"
          },
          "preferences": {
            "$ref": "#/components/schemas/OptimizationPreferences"
          },
          "professorConstraints": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProfessorConstraint"
            },
            "title": "Professorconstraints",
            "type": "array"
          },
          "requiredCourseCodes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Requiredcoursecodes",
            "type": "array"
          },
          "seed": {
            "default": 0,
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "selectedSectionIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Selectedsectionids",
            "type": "array"
          },
          "semester": {
            "default": "2026-1",
            "title": "Semester",
            "type": "string"
          },
          "targetCredits": {
            "anyOf": [
              {
                "maximum": 30.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetcredits"
          },
          "timeLimitSeconds": {
            "default": 3,
            "exclusiveMinimum": 0.0,
            "maximum": 8.0,
            "title": "Timelimitseconds",
            "type": "number"
          }
        },
        "required": [
          "datasetVersion"
        ],
        "title": "OptimizationCreate",
        "type": "object"
      },
      "OptimizationJobRead": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "cancelRequested": {
            "title": "Cancelrequested",
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "request": {
            "$ref": "#/components/schemas/OptimizationCreate"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/OptimizationJobStatus"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "request",
          "cancelRequested",
          "attempts",
          "createdAt",
          "updatedAt"
        ],
        "title": "OptimizationJobRead",
        "type": "object"
      },
      "OptimizationJobStatus": {
        "enum": [
          "QUEUED",
          "RUNNING",
          "OPTIMAL",
          "FEASIBLE",
          "INFEASIBLE",
          "TIME_LIMIT",
          "FAILED",
          "CANCELLED"
        ],
        "title": "OptimizationJobStatus",
        "type": "string"
      },
      "OptimizationPreferences": {
        "additionalProperties": false,
        "properties": {
          "avoidAfterMinute": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1440.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avoidafterminute"
          },
          "avoidBeforeMinute": {
            "anyOf": [
              {
                "exclusiveMaximum": 1440.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avoidbeforeminute"
          },
          "excludedDays": {
            "default": [],
            "items": {
              "enum": [
                "월",
                "화",
                "수",
                "목",
                "금",
                "토",
                "일"
              ],
              "type": "string"
            },
            "title": "Excludeddays",
            "type": "array"
          },
          "gapWeightPercent": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Gapweightpercent",
            "type": "integer"
          },
          "hardEndMinute": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1440.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardendminute"
          },
          "hardStartMinute": {
            "anyOf": [
              {
                "exclusiveMaximum": 1440.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardstartminute"
          },
          "maxDailyMinutes": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1440.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxdailyminutes"
          },
          "maxGapMinutes": {
            "anyOf": [
              {
                "maximum": 1440.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxgapminutes"
          },
          "minLunchMinutes": {
            "default": 0,
            "maximum": 150.0,
            "minimum": 0.0,
            "title": "Minlunchminutes",
            "type": "integer"
          },
          "minimizeCampusDays": {
            "default": true,
            "title": "Minimizecampusdays",
            "type": "boolean"
          },
          "minimizeChanges": {
            "default": true,
            "title": "Minimizechanges",
            "type": "boolean"
          },
          "minimizeGapMinutes": {
            "default": true,
            "title": "Minimizegapminutes",
            "type": "boolean"
          },
          "preferredDaysOff": {
            "default": [],
            "items": {
              "enum": [
                "월",
                "화",
                "수",
                "목",
                "금",
                "토",
                "일"
              ],
              "type": "string"
            },
            "title": "Preferreddaysoff",
            "type": "array"
          }
        },
        "title": "OptimizationPreferences",
        "type": "object"
      },
      "OptimizationResult": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OptimizationCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "reasons": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "solverVersion": {
            "title": "Solverversion",
            "type": "string"
          }
        },
        "required": [
          "solverVersion"
        ],
        "title": "OptimizationResult",
        "type": "object"
      },
...
      "ProfessorConstraint": {
        "additionalProperties": false,
        "properties": {
          "courseCode": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Coursecode",
            "type": "string"
          },
          "professor": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Professor",
            "type": "string"
          }
        },
        "required": [
          "courseCode",
          "professor"
        ],
        "title": "ProfessorConstraint",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PL-timeTabler API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/optimizations": {
      "post": {
        "operationId": "create_optimization_api_v1_optimizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationJobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Optimization",
        "tags": [
          "optimizations"
        ]
      }
    },
    "/api/v1/optimizations/compare": {
      "post": {
        "operationId": "compare_candidates_api_v1_optimizations_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Candidates",
        "tags": [
          "optimizations"
        ]
      }
    },
    "/api/v1/optimizations/{job_id}": {
      "delete": {
        "operationId": "cancel_optimization_api_v1_optimizations__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationJobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Optimization",
        "tags": [
          "optimizations"
        ]
      },
      "get": {
        "operationId": "get_optimization_api_v1_optimizations__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationJobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Optimization",
        "tags": [
          "optimizations"
        ]
      }
    },
...
  }
}